Dedicated Naming pattern for End-2-End tests
- clarify supported test-scenarios with explicit names that can be referred to in docs/discussions
