+++ smart-workflow-test/src_test/com/axonivy/utils/smart/workflow/model/openai/OpenAiModelE2E.java
@@ -1,4 +1,4 @@
-package com.axonivy.utils.smart.workflow.model.xai;
+package com.axonivy.utils.smart.workflow.model.openai;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 import com.axonivy.utils.smart.workflow.model.ChatModelFactory.AiConf;
-import com.axonivy.utils.smart.workflow.model.xai.internal.XAiServiceConnector;
+import com.axonivy.utils.smart.workflow.model.openai.internal.OpenAiServiceConnector.OpenAiConf;
 import com.axonivy.utils.smart.workflow.test.TestToolUserData;
 import com.axonivy.utils.smart.workflow.test.utils.TestUtils;
 
@@ -20,7 +20,7 @@ import ch.ivyteam.test.log.LoggerAccess;
 import dev.langchain4j.http.client.log.LoggingHttpClient;
 
 @IvyProcessTest
-class XAiModelIT {
+class OpenAiModelE2E {
 
   private static final BpmProcess AGENT_TOOLS = BpmProcess.name("TestToolUser");
 
@@ -29,8 +29,8 @@ class XAiModelIT {
 
   @BeforeEach
   void setup(AppFixture fixture) {
-    fixture.var(AiConf.DEFAULT_PROVIDER, XAiModelProvider.NAME); // enforce xAI!
-    fixture.var(XAiServiceConnector.XAiConf.API_KEY, TestUtils.getSystemProperty("XAI_API_KEY"));
+    fixture.var(AiConf.DEFAULT_PROVIDER, OpenAiModelProvider.NAME); // enforce OpenAI!
+    fixture.var(OpenAiConf.API_KEY, TestUtils.getSystemProperty("OPEN_AI_API_KEY"));
   }
 
   @Test
